+++ app/src/main/java/com/twilio/video/app/sdk/ConnectOptionsFactory.kt
@@ -120,14 +120,18 @@ class ConnectOptionsFactory(
             Preferences.AUDIO_AUTOMATIC_GAIN_CONTROL,
             Preferences.AUDIO_AUTOMATIC_GAIN_CONTROL_DEFAULT,
         )
-        val openSLESUsage = sharedPreferences.getBoolean(
-            Preferences.AUDIO_OPEN_SLES_USAGE,
-            Preferences.AUDIO_OPEN_SLES_USAGE_DEFAULT,
-        )
-        WebRtcAudioUtils.setWebRtcBasedAcousticEchoCanceler(!acousticEchoCanceler)
-        WebRtcAudioUtils.setWebRtcBasedNoiseSuppressor(!noiseSuppressor)
-        WebRtcAudioUtils.setWebRtcBasedAutomaticGainControl(!automaticGainControl)
-        WebRtcAudioManager.setBlacklistDeviceForOpenSLESUsage(!openSLESUsage)
+        // Removed due WebRTC-124 removing the WebRtcAudioUtils & WebRtcAudioManager classes
+        // Audio effects shall be added on a per-audio track basis. VBLOCKS-3744
+        //
+        // val openSLESUsage = sharedPreferences.getBoolean(
+        //     Preferences.AUDIO_OPEN_SLES_USAGE,
+        //     Preferences.AUDIO_OPEN_SLES_USAGE_DEFAULT,
+        // )
+
+        // WebRtcAudioUtils.setWebRtcBasedAcousticEchoCanceler(!acousticEchoCanceler)
+        // WebRtcAudioUtils.setWebRtcBasedNoiseSuppressor(!noiseSuppressor)
+        // WebRtcAudioUtils.setWebRtcBasedAutomaticGainControl(!automaticGainControl)
+        // WebRtcAudioManager.setBlacklistDeviceForOpenSLESUsage(!openSLESUsage)
 
         val isNetworkQualityEnabled = sharedPreferences.get(
             Preferences.ENABLE_NETWORK_QUALITY_LEVEL,